+++ NEWS.md
@@ -1,3 +1,5 @@
+# watcher (development version)
+
 # watcher 0.1.1
 
 * Updates bundled 'libfswatch' source package to 1.19.0-dev.