+++ NEWS.md
@@ -1,4 +1,4 @@
-# watcher (development version)
+# watcher 0.1.2
 
 * Adds `$get_path()` and `$is_running()` methods to the `Watcher` R6 class.
   + Use these rather than the fields `path` and `running`, as they have been made private.